requirements.txt
typer[all]>=0.9.0
langchain>=0.1.0
langchain-openai>=0.0.5
chromadb>=0.4.22
sentence-transformers>=2.2.2
tree-sitter>=0.20.4
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.20.1
tree-sitter-typescript>=0.20.3
openai>=1.12.0
pyyaml>=6.0.1
rich>=13.7.0
pytest>=7.4.4
pytest-cov>=4.1.0
black>=23.12.1
flake8>=7.0.0